+++ .gitignore
@@ -1,1 +1,1 @@
-packer_compiled.lua
+lazy-lock.json
