--- after/plugin/autopairs.lua
@@ -1,1 +1,0 @@
-require'nvim-autopairs'.setup()
