--- after/plugin/lsp.lua
@@ -1,128 +1,0 @@
-require("mason").setup({
-    ui = {
-        icons = {
-            package_installed = "✓",
-            package_pending = "➜",
-            package_uninstalled = "✗"
-        }
-    },
-    install_root_dir = os.getenv("NVIM_MASON_INSTALL") .. "/"
-})
-require("mason-lspconfig").setup()
-
-local lsp = require('lsp-zero')
-
-lsp.preset("recommended")
-
-lsp.ensure_installed({
-	'pylsp',
-	'clangd',
-    'lua_ls',
-    'cmake',
-    'jsonls'
-})
-
-require("lspconfig").cmake.setup{
-    filetypes = {'cmake', 'CMakeLists.txt'},
-}
-
-
-require("lspconfig").glsl_analyzer.setup{
-  filetypes = {'glsl', 'frag', 'vert', 'conf'} -- shaders are detected as conf for some reason (guessing #version <num> header does that)
-} -- https://github.com/nolanderc/glsl_analyzer
-
-require("lspconfig").pylsp.setup{
-    settings = {
-        pylsp = {
-            plugins = {
-                pycodestyle = {
-                    ignore = {"W391"},
-                    maxLineLength = 120
-                },
-                jedi = {
-                    environment = os.getenv("NVIM_PYTHON")
-                }
-            }
-        }
-    }
-}
-
-lsp.nvim_workspace()
-
-local has_words_before = function()
-  local line, col = unpack(vim.api.nvim_win_get_cursor(0))
-  return col ~= 0 and vim.api.nvim_buf_get_lines(0, line - 1, line, true)[1]:sub(col, col):match("%s") == nil
-end
-
-local cmp = require('cmp')
-local luasnip = require'luasnip'
-cmp.setup({
-  snippet = {
-    expand = function(args)
-      require('luasnip').lsp_expand(args.body)
-    end,
-  },
-  mapping = {
-    ["<Tab>"] = cmp.mapping(function(fallback)
-      if cmp.visible() then
-        cmp.select_next_item()
-      elseif luasnip.expand_or_jumpable() then
-        luasnip.expand_or_jump()
-      elseif has_words_before() then
-        cmp.complete()
-      else
-        fallback()
-      end
-    end, { "i", "s" }),
-    ["<S-Tab>"] = cmp.mapping(function(fallback)
-      if cmp.visible() then
-        cmp.select_prev_item()
-      elseif luasnip.jumpable(-1) then
-        luasnip.jump(-1)
-      else
-        fallback()
-      end
-    end, { "i", "s" }),
-    ['<C-j>'] = cmp.mapping(cmp.mapping.scroll_docs(-4), { 'i', 'c' }),
-    ['<C-k>'] = cmp.mapping(cmp.mapping.scroll_docs(4), { 'i', 'c' }),
-    ['<C-Space>'] = cmp.mapping(cmp.mapping.complete(), { 'i', 'c' }),
-    ['<C-y>'] = cmp.config.disable,
-    ['<C-e>'] = cmp.mapping({
-      i = cmp.mapping.abort(),
-      c = cmp.mapping.close(),
-    }),
-    ['<CR>'] = cmp.mapping.confirm({ select = true }),
-  },
-  sources = cmp.config.sources({
-    { name = 'nvim_lsp' },
-    { name = 'luasnip' },
-    { name = 'path' },
-    { name = 'buffer' },
-  })
-})
-
-lsp.set_preferences({
-    sign_icons = {
-        error = 'E',
-        warn = 'W',
-        hint = 'H',
-        info = 'I'
-    }
-})
-
-lsp.on_attach(function(client, bufnr)
-  local opts = {buffer = bufnr, remap = false}
-
-  vim.keymap.set("n", "gd", function() vim.lsp.buf.definition() end, opts)
-  vim.keymap.set("n", "K", function() vim.lsp.buf.hover() end, opts)
-  vim.keymap.set("n", "<leader>vws", function() vim.lsp.buf.workspace_symbol() end, opts)
-  vim.keymap.set("n", "<leader>vd", function() vim.diagnostic.open_float() end, opts)
-  vim.keymap.set("n", "<leader>nd", function() vim.diagnostic.goto_next() end, opts)
-  vim.keymap.set("n", "<leader>pd", function() vim.diagnostic.goto_prev() end, opts)
-  vim.keymap.set("n", "<leader>vca", function() vim.lsp.buf.code_action() end, opts)
-  vim.keymap.set("n", "<leader>vrr", function() vim.lsp.buf.references() end, opts)
-  vim.keymap.set("n", "<leader>vrn", function() vim.lsp.buf.rename() end, opts)
-  vim.keymap.set("n", "<leader>sh", function() vim.lsp.buf.signature_help() end, opts)
-end)
-
-lsp.setup()
