--- after/plugin/nvim-web-devicons.lua
@@ -1,1 +1,0 @@
-require('nvim-web-devicons').setup()
