--- after/plugin/treesitter.lua
@@ -1,29 +1,0 @@
-require'nvim-treesitter.configs'.setup {
-  -- A list of parser names, or "all" (the five listed parsers should always be installed)
-  ensure_installed = { "c", "lua", "vim", "vimdoc", "query", "glsl" },
-
-  -- Install parsers synchronously (only applied to `ensure_installed`)
-  sync_install = false,
-
-  -- Automatically install missing parsers when entering buffer
-  -- Recommendation: set to false if you don't have `tree-sitter` CLI installed locally
-  auto_install = true,
-
-  highlight = {
-    enable = true,
-
-    -- Setting this to true will run `:h syntax` and tree-sitter at the same time.
-    -- Set this to `true` if you depend on 'syntax' being enabled (like for indentation).
-    -- Using this option may slow down your editor, and you may see some duplicate highlights.
-    -- Instead of true it can also be a list of languages
-    additional_vim_regex_highlighting = false,
-  },
-}
-
-require'treesitter-context'.setup()
-
-vim.api.nvim_set_hl(0, "TreeSitterContext", {bg = "none"})
-vim.api.nvim_set_hl(0, "TreeSitterContextBottom", {bg = "none", underline=true})
-vim.api.nvim_set_hl(0, "TreeSitterContextLineNumber", {fg = "#ffffff"})
-
-vim.treesitter.language.register('glsl', {'vert', 'frag', 'conf'}) -- shaders are detected as conf for some reason (guessing #version <num> header does that)
