+++ lua/meepowned/init.lua
@@ -1,2 +1,24 @@
-require("meepowned.remap")
 require("meepowned.set")
+require("meepowned.remap")
+require("meepowned.lazy_init")
+
+local augroup = vim.api.nvim_create_augroup
+local MeepOwnedGroup = augroup('MeepOwned', {})
+local autocmd = vim.api.nvim_create_autocmd
+
+autocmd('LspAttach', {
+    callback = function(e)
+        local opts = { buffer = e.buf }
+
+        vim.keymap.set("n", "gd", function() vim.lsp.buf.definition() end, opts)
+        vim.keymap.set("n", "K", function() vim.lsp.buf.hover() end, opts)
+        vim.keymap.set("n", "<leader>vws", function() vim.lsp.buf.workspace_symbol() end, opts)
+        vim.keymap.set("n", "<leader>vd", function() vim.diagnostic.open_float() end, opts)
+        vim.keymap.set("n", "<leader>nd", function() vim.diagnostic.goto_next() end, opts)
+        vim.keymap.set("n", "<leader>pd", function() vim.diagnostic.goto_prev() end, opts)
+        vim.keymap.set("n", "<leader>vca", function() vim.lsp.buf.code_action() end, opts)
+        vim.keymap.set("n", "<leader>vrr", function() vim.lsp.buf.references() end, opts)
+        vim.keymap.set("n", "<leader>vrn", function() vim.lsp.buf.rename() end, opts)
+        vim.keymap.set("n", "<leader>sh", function() vim.lsp.buf.signature_help() end, opts)
+    end
+})
