+++ lua/meepowned/lazy/fugitive.lua
@@ -1,0 +1,8 @@
+return {
+    {
+        'tpope/vim-fugitive',
+        config = function()
+            vim.keymap.set("n", "<leader>gs", vim.cmd.Git)
+        end
+    },
+}
