+++ lua/meepowned/lazy/lualine.lua
@@ -1,0 +1,34 @@
+return {
+    {
+        'nvim-lualine/lualine.nvim',
+        dependencies = {
+            'nvim-tree/nvim-web-devicons',
+            'navarasu/onedark.nvim',
+        },
+        config = function()
+            require('lualine').setup({
+                options = {
+                    icons_enabled = true,
+                    theme = 'onedark'
+                },
+                sections = {
+                    lualine_a = { 'mode', 'filename' },
+                    lualine_b = { 'branch', 'diff', 'diagnostics' },
+                    lualine_c = { 'hostname' },
+                    lualine_x = { 'selectioncount', 'encoding', 'fileformat' },
+                    lualine_y = { 'filetype' },
+                    lualine_z = { 'progress', 'location' }
+
+                },
+                inactive_sections = {
+                    lualine_a = {},
+                    lualine_b = { 'filename' },
+                    lualine_c = { 'hostname' },
+                    lualine_x = { 'fileformat' },
+                    lualine_y = { 'filetype' },
+                    lualine_z = {}
+                }
+            })
+        end
+    },
+}
