+++ lua/meepowned/lazy/nvim-web-devicons.lua
@@ -1,0 +1,8 @@
+return {
+    {
+        'nvim-tree/nvim-web-devicons',
+        config = function()
+            require('nvim-web-devicons').setup({})
+        end
+    },
+}
