+++ lua/meepowned/lazy/telescope.lua
@@ -1,0 +1,26 @@
+return {
+    {
+        'nvim-telescope/telescope.nvim',
+        tag = "0.1.5",
+        dependencies = {
+            'nvim-lua/plenary.nvim',
+        },
+
+        config = function()
+            require('telescope').setup({})
+
+            local builtin = require('telescope.builtin')
+
+            vim.keymap.set('n', '<leader>?', require('telescope.builtin').oldfiles,
+                { desc = '[?] Find recently opened files' })
+            vim.keymap.set('n', '<leader>gf', require('telescope.builtin').git_files, { desc = 'Search [G]it [F]iles' })
+            vim.keymap.set('n', '<leader>sf', require('telescope.builtin').find_files, { desc = '[S]earch [F]iles' })
+            -- vim.keymap.set('n', '<leader>sh', require('telescope.builtin').help_tags, { desc = '[S]earch [H]elp' })
+            vim.keymap.set('n', '<leader>sw', require('telescope.builtin').grep_string,
+                { desc = '[S]earch current [W]ord' })
+            vim.keymap.set('n', '<leader>sg', require('telescope.builtin').live_grep, { desc = '[S]earch by [G]rep' })
+            vim.keymap.set('n', '<leader>sd', require('telescope.builtin').diagnostics,
+                { desc = '[S]earch [D]iagnostics' })
+        end
+    },
+}
