+++ lua/meepowned/lazy/undotree.lua
@@ -1,0 +1,8 @@
+return {
+    {
+        'mbbill/undotree',
+        config = function()
+            vim.keymap.set("n", "<leader>u", vim.cmd.UndotreeToggle)
+        end
+    },
+}
